+++ R/addKernelDensity.R
@@ -89,7 +89,7 @@ addKernelDensity <- function(pump.subset = "pooled", pump.select = NULL,
 
       } else if (pump.subset == "individual") {
         if (neighborhood.type == "walking") {
-          n.data <- neighborhoodWalking(latlong = latlong, multi.core = cores)
+          n.data <- walkingB(latlong = latlong, multi.core = cores)
           cases <- pumpCase(n.data)
 
         } else if (neighborhood.type == "voronoi") {
@@ -129,20 +129,29 @@ addKernelDensity <- function(pump.subset = "pooled", pump.select = NULL,
 
     } else if (all(is.numeric(pump.subset))) {
       if (neighborhood.type == "walking") {
-        n.data <- neighborhoodWalking(latlong = latlong, multi.core = cores)
-        obs.neighborhood <- as.numeric(names(n.data$paths))
-
-        if (any(abs(pump.subset) %in% obs.neighborhood == FALSE)) {
-          stop('For walking neighborhoods, only 3 through 12 are valid.')
-        }
-
+        n.data <- walkingB(latlong = latlong, multi.core = cores)
         cases.list <- pumpCase(n.data)
 
+        p.nm <- names(cases.list)
+        p.obs <- as.integer(substr(p.nm, 2, nchar(p.nm)))
+
         if (all(pump.subset > 0)) {
-          cases <- cases.list[paste0("p", pump.subset)]
+          if (all(pump.subset %in% p.obs)) {
+            cases <- cases.list[paste0("p", pump.subset)]
+          } else {
+            stop('pump.subset valid only for pumps 3 through 12.',
+              call. = FALSE)
+          }
         } else if (all(pump.subset < 0)) {
-          sel <- names(cases.list) %in% paste0("p", abs(pump.subset)) == FALSE
-          cases <- cases.list[sel]
+          neg.pump <- paste0("p", abs(pump.subset))
+
+          if (all(neg.pump %in% p.nm)) {
+            sel <- !names(cases.list) %in% paste0("p", abs(pump.subset))
+            cases <- cases.list[sel]
+          } else {
+            stop('pump.subset valid only for pumps 3 through 12.',
+              call. = FALSE)
+          }
         } else {
           stop("Use all positive or all negative numbers for pump.subset.")
         }
@@ -183,7 +192,7 @@ addKernelDensity <- function(pump.subset = "pooled", pump.select = NULL,
     }
 
   } else {
-    n.data <- neighborhoodWalking(pump.select, latlong = latlong,
+    n.data <- walkingB(pump.select, latlong = latlong,
       multi.core = cores)
     cases <- pumpCase(n.data)