+++ NEWS.md
@@ -1,3 +1,9 @@
+### cholera 0.8.0.9579
+
+- use walkingB() in addKernelDensity().
+- restrict pump.subset to observed pump neighborhoods.
+
+
 ### cholera 0.8.0.9578
 
 - use walkingB() in orientationPlot().